fix(rpc/v07): remove dto leftover type for `ResourceBounds`
