+++ crates/rpc/src/method/add_declare_transaction.rs
@@ -365,20 +365,9 @@ pub async fn add_declare_transaction(
                     add_transaction::Declare::V3(add_transaction::DeclareV3 {
                         signature: tx.signature,
                         nonce: tx.nonce,
-                        nonce_data_availability_mode:
-                            pathfinder_common::transaction::DataAvailabilityMode::from(
-                                tx.nonce_data_availability_mode,
-                            )
-                            .into(),
-                        fee_data_availability_mode:
-                            pathfinder_common::transaction::DataAvailabilityMode::from(
-                                tx.fee_data_availability_mode,
-                            )
-                            .into(),
-                        resource_bounds: pathfinder_common::transaction::ResourceBounds::from(
-                            tx.resource_bounds,
-                        )
-                        .into(),
+                        nonce_data_availability_mode: tx.nonce_data_availability_mode.into(),
+                        fee_data_availability_mode: tx.fee_data_availability_mode.into(),
+                        resource_bounds: tx.resource_bounds.into(),
                         tip: tx.tip,
                         paymaster_data: tx.paymaster_data,
                         contract_class: contract_definition,
@@ -415,6 +404,7 @@ mod tests {
     use std::sync::LazyLock;
 
     use pathfinder_common::macro_prelude::*;
+    use pathfinder_common::transaction::{DataAvailabilityMode, ResourceBound, ResourceBounds};
     use pathfinder_common::{
         CasmHash,
         ContractAddress,
@@ -438,14 +428,7 @@ mod tests {
         BroadcastedDeclareTransactionV2,
         BroadcastedDeclareTransactionV3,
     };
-    use crate::types::{
-        CairoContractClass,
-        ContractClass,
-        DataAvailabilityMode,
-        ResourceBound,
-        ResourceBounds,
-        SierraContractClass,
-    };
+    use crate::types::{CairoContractClass, ContractClass, SierraContractClass};
 
     pub static CONTRACT_CLASS: LazyLock<CairoContractClass> = LazyLock::new(|| {
         ContractClass::from_definition_bytes(CONTRACT_DEFINITION)
